requirements.txt
python-decouple
requests
transitions
pandas